+++ lib/API/Actions/HiddenContent.php
@@ -15,7 +15,7 @@ class HiddenContent {
 	/**
 	 * The namespace.
 	 *
-	 * @since Unreleased
+	 * @since 0.5.13
 	 * @var   string
 	 */
 	protected $namespace;