+++ lib/API/Options.php
@@ -15,7 +15,7 @@ class Options {
 	/**
 	 * The namespace.
 	 *
-	 * @since Unreleased
+	 * @since 0.5.13
 	 * @var string
 	 */
 	protected $namespace;