+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.13] - 2025-05-12
+
 ### Added
 
 - Adds missing class property declarations (PHP 8.2 support).
@@ -578,7 +580,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.5.12...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.5.13...HEAD
+[0.5.13]: https://github.com/26b/unbabble/compare/0.5.12...0.5.13
 [0.5.12]: https://github.com/26b/unbabble/compare/0.5.11...0.5.12
 [0.5.11]: https://github.com/26b/unbabble/compare/0.5.10...0.5.11
 [0.5.10]: https://github.com/26b/unbabble/compare/0.5.9...0.5.10